- Apply UpperCamelCase naming convention across files
- Rename base.ts → Fixtures.ts

- Move API-related code to new API_Helper.ts file

- Add blockAds and saveAdditionsAttachments functions for better clarity and maintainability

- Add scrollToElement and waitForImageToLoad functions to BasePage.ts

- Update all affected imports and references
