+++ Helper/Fixtures.ts
@@ -11,9 +11,9 @@ import { ProductsPage } from '../pages/ProductsPage';
 import { ProductPage } from '../pages/ProductPage';
 import { CartPage } from '../pages/CartPage';
 import { PaymentPage } from '../pages/PaymentPage';
-import { AuthorizationAPI } from '../API/authorization';
-import { ProductsAPI } from '../API/products';
-import { randomUUID } from 'crypto';
+import { AuthorizationAPI } from '../API/Authorization';
+import { ProductsAPI } from '../API/Products.ts';
+import { blockAds, saveAdditionsAttachments } from './Tools.ts';
 import dotenv from 'dotenv';
 dotenv.config();
 
@@ -93,34 +93,17 @@ export const test = baseTest.extend<MyFixtures>({
     // Thanks to Checkly for that https://www.youtube.com/watch?v=hegZS46J0rA
     // -----
     saveData: [async ({ context, page }, use, testInfo) => {
-        await context.route("**/*", route => {
-          route.request().url().startsWith("https://googleads.") ?
-          route.abort() : route.continue();
-          return;
-        })
+        await blockAds(context);
 
         // Start tracing (with screenshots, snapshots, and sources)
         await context.tracing.start({ screenshots: true, snapshots: true, sources: true });
 
         await use();
+
         testInfo.setTimeout(30000);
 
         try{
-            const testTitle = testInfo.title.replace(/[^a-zA-Z0-9-_]/g, '_');
-            // Stop tracing and save it
-            const tracePath = `test-resultsSave/traces/trace-${testTitle}.zip`;
-            await context.tracing.stop({ path: tracePath });
-            testInfo.annotations.push({ type: 'testrail_attachment', description: tracePath });
-            
-            let screenshotPath = `test-resultsSave/screenshots/screenshot-${testTitle}.png`;
-            const videoPath = `test-resultsSave/videos/video-${testTitle}.webm`;
-            await page.screenshot({ path: screenshotPath, fullPage: true });
-            await page.close();
-            testInfo.annotations.push({ type: 'testrail_attachment', description: screenshotPath });
-            if (page.video()) {
-                await page.video()?.saveAs(videoPath);
-                testInfo.annotations.push({ type: 'testrail_attachment', description: videoPath });
-            }
+            await saveAdditionsAttachments(context, page, testInfo);
         }catch(err){
             console.error("Error during saving attachments:", err);
         }
@@ -130,28 +113,3 @@ export const test = baseTest.extend<MyFixtures>({
 });
 
 export {expect} from '@playwright/test';
-
-export const Status = { success: 200, resourceCreated: 201, badReq: 400, notFound: 404, methodNotAllowed: 405, serverError: 500};
-export const Methods = { GET: 'GET', POST: 'POST', PUT: 'PUT', DELETE: 'DELETE'};
-
-export const APIEndPoints = {
-    verifyLogin: '/api/verifyLogin',
-    createAccount: '/api/createAccount',
-    deleteAccount: '/api/deleteAccount',
-    getUserDetailByEmail: '/api/getUserDetailByEmail',
-    brandList: '/api/brandsList',
-    searchProduct: '/api/searchProduct',
-};
-
-export const Messages = {
-    userFoundMessage: "User exists!",
-    userNotFoundMessage: "User not found!",
-    userCreatedMessage: "User created!",
-    userDeletedMessage: "Account deleted!",
-    accountNotFoundMessage: "Account not found with this email, try another email!",
-    badRequestMessage: (method: string) => `Bad request, email or password parameter is missing in ${method} request.`,
-    badRequestParameterMessage: (method: string, parameter: string) => `Bad request, ${parameter} parameter is missing in ${method} request.`,
-    notSupportedReqMethodMessage: "This request method is not supported.",
-    emailAlreadyExistsMessage: "Email already exists!",
-    methodNotAllowedMessage: (method: string) => `Method \"${method}\" not allowed.`,
-}
