+++ tests/E2E/product_catalog.spec.ts
@@ -1,5 +1,5 @@
-import { test } from '../../Helper/base.ts';
-import { getEnv } from '../../Helper/tools.ts';
+import { test } from '../../Helper/Fixtures.ts';
+import { getEnv } from '../../Helper/Tools.ts';
 
 test.describe("E2E Product & Catalog", () =>{
   test.beforeEach(async ({ homePage }) =>{
